MDL-63296 profile: display phone1, phone2 and other fields in profile
Sets phone1,phone2,institution,department,idnumber,address as
default selected hiddenuserfields in the "User policies" config and
modify de profile.php in order to control its visibility with this
config instead to use showuseridentity which should just control
visibility fields in user lists and not the profile. Admins are
required to unselect this fields in order to display them in profile.

Includes a db patch in lib/db/upgrade.php in order to add this fields
in the hiddenuserfields config in upgraded installations to ensure
this fields stay hidden until an administrator unselect them.